Move the default code into the correct package
